remove ssr routes in favor of query params
