+++ db/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "db"
-version = "0.1.0"
+version = "0.1.1"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
