+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fieldz-desktop",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "scripts": {
     "preinstall": "cd webview && npm install"
   }
