--- webview/src/routes/region/[id]/+page.ts
@@ -1,5 +1,0 @@
-export function load({ params }) {
-	return {
-		id: params.id
-	};
-}
