+++ webview/src/routes/regions/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+	import RegionList from './RegionList.svelte';
+</script>
+
+<RegionList />
